+++ src/demo/java/Todo.java
@@ -1,0 +1,6 @@
+class Todo {
+
+    String id;
+    String text;
+    boolean completed;
+}