+++ src/demo/java/TodoRepository.java
@@ -1,0 +1,10 @@
+import java.util.List;
+
+import static java.util.Collections.emptyList;
+
+class TodoRepository {
+
+    List<Todo> todos() {
+        return emptyList();
+    }
+}